requirements.txt
Django>=5.2.1
Pillow>=10.0.0
django-debug-toolbar>=4.2.0
reportlab>=4.0.0
djangorestframework>=3.14.0
django-simple-history>=3.4.0
django-import-export>=3.3.0
django-filter>=23.0 